[Haml] Automatically filter all plain-text blocks through the filter specified with the --autofilter command-line option.
This is useful for hand-crafted haml pages where I want *all* content to be
markdown-formatted. If using something like markdown, make sure you don't
nest a plain-text block under a %p tag, or you'll generate nonsensical markup.

This has *not* been tested extensively -- indeed I can't run the unit tests
because `rake test` asks for a non-existent sass gem. Do *not* use this in its
current state for dynamically-generated content, and always check the output. I
am not terribly confident in the quality of this change because I don't really
follow the dozens of state variables in parser.rb so I just hacked around until
it worked for me.